pyproject.toml
[project]
name = "rebelist-revelations"
version = "0.11.0"
description = "An experimental RAG application that scans Confluence for instant answers from your internal documentation."
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "atlassian-python-api>=4.0.7,<5.0.0",
    "dependency-injector>=4.47.1,<5.0.0",
    "docling>=2.66.0,<3.0.0",
    "fastembed>=0.7.4",
    "huggingface-hub>=0.36.0,<2.0.0",
    "langchain>=1.1.3,<2.0.0",
    "langchain-community>=0.4.1,<1.0.0",
    "langchain-core>=1.0.4,<2.0.0",
    "langchain-ollama>=1.0.0,<2.0.0",
    "langchain-qdrant>=1.1.0,<2.0.0",
    "langchain-text-splitters>=1.0.0,<2.0.0",
    "langchainhub>=0.1.21,<1.0.0",
    "loguru>=0.7.3,<1.0.0",
    "prompt-toolkit>=3.0.52",
    "pydantic>=2.12.5,<3.0.0",
    "pydantic-settings>=2.12.0,<3.0.0",
    "pymongo>=4.13.0,<5.0.0",
    "python-dotenv>=1.1.0,<2.0.0",
    "qdrant-client>=1.16.1,<2.0.0",
    "rich-click>=1.8.9,<2.0.0",
    "sentence-transformers>=5.1.2,<6.0.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.3.0,<5.0.0",
    "pyright>=1.1.405,<2.0.0",
    "pytest>=9.0.0,<10.0.0",
    "pytest-cov>=7.0.0,<8.0.0",
    "pytest-mock>=3.15.0,<4.0.0",
    "ruff>=0.14.7,<1.0.0",
    "types-mock>=5.2.0.20250809,<6.0.0"
]

[project.scripts]
revelations = "rebelist.revelations.main:console"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rebelist"]

### pyright ###

[tool.pyright]
venvPath = "."
venv = ".venv"
include = ["src", "tests"]
typeCheckingMode = "strict"
reportMissingTypeStubs = false
reportUnknownMemberType = false
reportUnknownLambdaType = false
exclude = [
    "**/__pycache__"
]

### ruff ###

[tool.ruff]
line-length = 120

[tool.ruff.format]
quote-style = "single"
docstring-code-format = true

[tool.ruff.lint]
select = [
    "A", # Aesthetic rules (e.g., blank lines, comment formatting)
    "E", # Errors (e.g., syntax errors, indentation issues)
    "W", # Warnings (e.g., line length, trailing whitespace)
    "F", # Pyflakes (e.g., unused imports, undefined variables)
    "I", # Import rules (e.g., import order, unused imports)
    "B", # Bugbear (e.g., potential bugs, mutable default arguments)
    "C4", # McCabe complexity (e.g., function complexity)
    "N", # Naming conventions (e.g., PEP8 naming standards)
    "D", # Docstrings (e.g., missing docstrings, formatting issues)
    "RUF100", # Mistakes (e.g., common mistakes or poor patterns)
]
ignore = [
    "ANN", # Disable all type annotation-related checks
    "PYI", # Disable checks for .pyi stub files
    "PLR", # Disable Pylint rules related to typing
    "D107", # Disable documentation on __init__ methods
    "D100", # Disable missing docstring in public module
    "D104", # Disable missing docstring in public package
    "D101", # Disable undocumented public class definitions.
]

[tool.ruff.lint.pydocstyle]
convention = "google"

### pytest ###

[tool.pytest]
pythonpath = ["src"]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = ["--strict-markers", "--strict-config", "--import-mode=importlib"]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "e2e: End-to-end tests"
]
filterwarnings = [
    "ignore:builtin type SwigPyPacked:DeprecationWarning",
    "ignore:builtin type SwigPyObject:DeprecationWarning",
    "ignore:builtin type swigvarlink:DeprecationWarning",
]

### coverage ###

[tool.coverage.run]
branch = true
source = ["src"]
omit = [
    "*/site-packages/*",
    "*/dependency_injector/*",
    "**/__init__.py",
]

[tool.coverage.report]
show_missing = true
skip_covered = true

[tool.coverage.html]
directory = "var/coverage"
